requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
logging
azure-storage-blob
PyPDF2
msrest
azure-cognitiveservices-vision-computervision
openai
pandas
pyodbc